requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
requests>=2.31.0
pydantic>=2.5.0
python-multipart>=0.0.6
solana>=0.30.2
solders>=0.19.0
httpx>=0.23.0,<0.24.0
python-dotenv>=1.0.0
base58>=2.1.1
construct>=2.10.68
